pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "art-scripts"
version = "1.0.2"
description = ""
license = { text="MIT" }
authors = [ { name="cromachina" } ]
requires-python = ">=3.12"
dependencies = [
    "ffmpeg-python",
    "pillow",
]

[project.scripts]
art-archive = "archive:main"
copy-to-archive = "copy_to_archive:main"
hash-rename = "hash_rename:main"
del-frames = "del_frames:main"
